+++ android/src/main/java/io/openim/flutter_openim_sdk/FlutterOpenimSdkPlugin.java
@@ -20,6 +20,7 @@ import io.openim.flutter_openim_sdk.manager.IMManager;
 import io.openim.flutter_openim_sdk.manager.MessageManager;
 import io.openim.flutter_openim_sdk.manager.SignalingManager;
 import io.openim.flutter_openim_sdk.manager.UserManager;
+import io.openim.flutter_openim_sdk.manager.WorkMomentsManager;
 
 
 /**
@@ -39,6 +40,7 @@ public class FlutterOpenimSdkPlugin implements FlutterPlugin, MethodCallHandler 
     private static ConversationManager conversationManager;
     private static GroupManager groupManager;
     private static SignalingManager signalingManager;
+    private static WorkMomentsManager workMomentsManager;
 
     public FlutterOpenimSdkPlugin() {
     }
@@ -51,6 +53,7 @@ public class FlutterOpenimSdkPlugin implements FlutterPlugin, MethodCallHandler 
         FlutterOpenimSdkPlugin.conversationManager = new ConversationManager();
         FlutterOpenimSdkPlugin.groupManager = new GroupManager();
         FlutterOpenimSdkPlugin.signalingManager = new SignalingManager();
+        FlutterOpenimSdkPlugin.workMomentsManager = new WorkMomentsManager();
     }
 
     @Override