+++ ios/Classes/SwiftFlutterOpenimSdkPlugin.swift
@@ -9,6 +9,7 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
     let groupManager: GroupManager
     let userManger: UserManager
     let signalingManager: SignalingManager
+    let workMomentsManager: WorkMomentsManager
     
     init(channel: FlutterMethodChannel) {
         self.imManager = IMMananger(channel: channel)
@@ -18,6 +19,7 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
         self.groupManager = GroupManager(channel: channel)
         self.userManger = UserManager(channel: channel)
         self.signalingManager = SignalingManager(channel: channel)
+        self.workMomentsManager = WorkMomentsManager(channel: channel)
     }
     
     public static func register(with registrar: FlutterPluginRegistrar) {
@@ -43,6 +45,8 @@ public class SwiftFlutterOpenimSdkPlugin: NSObject, FlutterPlugin {
             userManger.handleMethod(call: call, result: result)
         case "signalingManager":
             signalingManager.handleMethod(call: call, result: result)
+        case "workMomentsManager":
+            workMomentsManager.handleMethod(call: call, result: result)
         default:
             print("Handle ManagerName Error: \(managerName) not found")
         }