+++ ios/Framework/OpenIMCore.framework/Headers/Open_im_sdk_callback.objc.h
@@ -23,10 +23,14 @@
 @class Open_im_sdk_callbackOnFriendshipListener;
 @protocol Open_im_sdk_callbackOnGroupListener;
 @class Open_im_sdk_callbackOnGroupListener;
+@protocol Open_im_sdk_callbackOnOrganizationListener;
+@class Open_im_sdk_callbackOnOrganizationListener;
 @protocol Open_im_sdk_callbackOnSignalingListener;
 @class Open_im_sdk_callbackOnSignalingListener;
 @protocol Open_im_sdk_callbackOnUserListener;
 @class Open_im_sdk_callbackOnUserListener;
+@protocol Open_im_sdk_callbackOnWorkMomentsListener;
+@class Open_im_sdk_callbackOnWorkMomentsListener;
 @protocol Open_im_sdk_callbackSendMsgCallBack;
 @class Open_im_sdk_callbackSendMsgCallBack;
 
@@ -38,7 +42,7 @@
 @protocol Open_im_sdk_callbackOnAdvancedMsgListener <NSObject>
 - (void)onRecvC2CReadReceipt:(NSString* _Nullable)msgReceiptList;
 - (void)onRecvGroupReadReceipt:(NSString* _Nullable)groupMsgReceiptList;
-- (void)onRecvMessageRevoked:(NSString* _Nullable)msgId;
+- (void)onRecvMessageRevoked:(NSString* _Nullable)msgID;
 - (void)onRecvNewMessage:(NSString* _Nullable)message;
 @end
 
@@ -84,6 +88,10 @@
 - (void)onJoinedGroupDeleted:(NSString* _Nullable)groupInfo;
 @end
 
+@protocol Open_im_sdk_callbackOnOrganizationListener <NSObject>
+- (void)onOrganizationUpdated;
+@end
+
 @protocol Open_im_sdk_callbackOnSignalingListener <NSObject>
 - (void)onInvitationCancelled:(NSString* _Nullable)invitationCancelledCallback;
 - (void)onInvitationTimeout:(NSString* _Nullable)invitationTimeoutCallback;
@@ -96,6 +104,10 @@
 
 @protocol Open_im_sdk_callbackOnUserListener <NSObject>
 - (void)onSelfInfoUpdated:(NSString* _Nullable)userInfo;
+@end
+
+@protocol Open_im_sdk_callbackOnWorkMomentsListener <NSObject>
+- (void)onRecvNewNotification;
 @end
 
 @protocol Open_im_sdk_callbackSendMsgCallBack <NSObject>
@@ -116,9 +128,13 @@
 
 @class Open_im_sdk_callbackOnGroupListener;
 
+@class Open_im_sdk_callbackOnOrganizationListener;
+
 @class Open_im_sdk_callbackOnSignalingListener;
 
 @class Open_im_sdk_callbackOnUserListener;
+
+@class Open_im_sdk_callbackOnWorkMomentsListener;
 
 @class Open_im_sdk_callbackSendMsgCallBack;
 
@@ -138,7 +154,7 @@
 - (nonnull instancetype)initWithRef:(_Nonnull id)ref;
 - (void)onRecvC2CReadReceipt:(NSString* _Nullable)msgReceiptList;
 - (void)onRecvGroupReadReceipt:(NSString* _Nullable)groupMsgReceiptList;
-- (void)onRecvMessageRevoked:(NSString* _Nullable)msgId;
+- (void)onRecvMessageRevoked:(NSString* _Nullable)msgID;
 - (void)onRecvNewMessage:(NSString* _Nullable)message;
 @end
 
@@ -200,6 +216,14 @@
 - (void)onJoinedGroupDeleted:(NSString* _Nullable)groupInfo;
 @end
 
+@interface Open_im_sdk_callbackOnOrganizationListener : NSObject <goSeqRefInterface, Open_im_sdk_callbackOnOrganizationListener> {
+}
+@property(strong, readonly) _Nonnull id _ref;
+
+- (nonnull instancetype)initWithRef:(_Nonnull id)ref;
+- (void)onOrganizationUpdated;
+@end
+
 @interface Open_im_sdk_callbackOnSignalingListener : NSObject <goSeqRefInterface, Open_im_sdk_callbackOnSignalingListener> {
 }
 @property(strong, readonly) _Nonnull id _ref;
@@ -220,6 +244,14 @@
 
 - (nonnull instancetype)initWithRef:(_Nonnull id)ref;
 - (void)onSelfInfoUpdated:(NSString* _Nullable)userInfo;
+@end
+
+@interface Open_im_sdk_callbackOnWorkMomentsListener : NSObject <goSeqRefInterface, Open_im_sdk_callbackOnWorkMomentsListener> {
+}
+@property(strong, readonly) _Nonnull id _ref;
+
+- (nonnull instancetype)initWithRef:(_Nonnull id)ref;
+- (void)onRecvNewNotification;
 @end
 
 @interface Open_im_sdk_callbackSendMsgCallBack : NSObject <goSeqRefInterface, Open_im_sdk_callbackSendMsgCallBack> {