+++ lib/src/enum/group_at_type.dart
@@ -1,0 +1,6 @@
+class GroupAtType {
+  static const atNormal = 0;
+  static const atMe = 1;
+  static const atAll = 2;
+  static const atAllAtMe = 3;
+}