+++ lib/src/listener/workmoments_listener.dart
@@ -1,0 +1,9 @@
+class OnWorkMomentsListener {
+  Function()? onRecvNewNotification;
+
+  OnWorkMomentsListener({this.onRecvNewNotification});
+
+  void recvNewNotification() {
+    onRecvNewNotification?.call();
+  }
+}